+++ apps/web/src/api/scenarios/fetch-scenario-by-id.tsx
