+++ apps/web/src/api/scenarios/fetch-scenarios.tsx
